RtCW 64 bit fixes
Had to change the VM API in order to pass intptr_t instead of int as
well as various other pointer messing.

+++ src/botlib/be_ai_goal.c
@@ -35,6 +35,8 @@ If you have questions concerning this license or the applicable additional terms
  *
  *****************************************************************************/
 
+#include <stdint.h>
+
 #include "../game/q_shared.h"
 #include "l_utils.h"
 #include "l_libvar.h"
@@ -137,7 +139,7 @@ typedef struct iteminfo_s
 	int number;                         //number of the item info
 } iteminfo_t;
 
-#define ITEMINFO_OFS( x ) (int)&( ( (iteminfo_t *)0 )->x )
+#define ITEMINFO_OFS( x ) (size_t)&( ( (iteminfo_t *)0 )->x )
 
 fielddef_t iteminfo_fields[] =
 {
