+++ src/botlib/be_ai_weap.c
@@ -35,6 +35,8 @@ If you have questions concerning this license or the applicable additional terms
  *
  *****************************************************************************/
 
+#include <stdint.h>
+
 #include "../game/q_shared.h"
 #include "l_libvar.h"
 #include "l_log.h"
@@ -54,8 +56,8 @@ If you have questions concerning this license or the applicable additional terms
 //#define DEBUG_AI_WEAP
 
 //structure field offsets
-#define WEAPON_OFS( x ) (int)&( ( (weaponinfo_t *)0 )->x )
-#define PROJECTILE_OFS( x ) (int)&( ( (projectileinfo_t *)0 )->x )
+#define WEAPON_OFS( x ) (size_t)&( ( (weaponinfo_t *)0 )->x )
+#define PROJECTILE_OFS( x ) (size_t)&( ( (projectileinfo_t *)0 )->x )
 
 //weapon definition
 fielddef_t weaponinfo_fields[] =
