+++ src/qcommon/vm_local.h
@@ -28,6 +28,7 @@ If you have questions concerning this license or the applicable additional terms
 
 #include "../game/q_shared.h"
 #include "qcommon.h"
+#include <stdint.h>
 
 typedef enum {
 	OP_UNDEF,
@@ -118,7 +119,7 @@ typedef enum {
 
 
 
-typedef int vmptr_t;
+typedef intptr_t vmptr_t;
 
 typedef struct vmSymbol_s {
 	struct vmSymbol_s   *next;
@@ -134,7 +135,7 @@ struct vm_s {
 	// DO NOT MOVE OR CHANGE THESE WITHOUT CHANGING THE VM_OFFSET_* DEFINES
 	// USED BY THE ASM CODE
 	int programStack;               // the vm may be recursively entered
-	int ( *systemCall )( int *parms );
+	intptr_t ( *systemCall )( intptr_t *parms );
 
 	//------------------------------------
 
@@ -142,7 +143,7 @@ struct vm_s {
 
 	// for dynamic linked modules
 	void        *dllHandle;
-	int ( QDECL *entryPoint )( int callNum, ... );
+	intptr_t ( QDECL *entryPoint )( intptr_t callNum, ... );
 
 	// for interpreted modules
 	qboolean currentlyInterpreting;
